revise editions work
