+++ app/assets/stylesheets/application/heliotrope.scss
@@ -541,9 +541,13 @@ footer.press {
 
   .previous-edition-metrics {
     // the Altmetric widgets are always the same width
-    max-width: 110px;
+    max-width: fit-content;
     padding-top: .5em;
     border-top: 2px solid #d5d5d5;
+    font-size: 13px;
+    .pub-date-edition-name {
+      margin-bottom: .5em;
+    }
   }
   .fulcrum_sidebar {
     // In hyrax2 this defaults to grey which we don't want
