+++ config/locales/en.yml
@@ -295,9 +295,9 @@ en:
         isbn: "Enter as many ISBN values as you want. Put the format in parentheses. Label the format according to the following rules: 123-123456-1 (hardcover), 123-123459-4 (paper), 123-123457-2 (ebook), 123-123458-3 (open access)"
         language: "The language of the monograph's content."
         location: "The place name used in auto-generated citations."
-        next_edition: "URL pointing to the next edition of the Monograph"
+        next_edition: "Full DOI (or other) URL pointing to the next edition of the Monograph"
         press: "The monograph's publisher/press on Fulcrum."
-        previous_edition: "URL pointing to the previous edition of the Monograph"
+        previous_edition: "Full DOI (or other) URL pointing to the previous edition of the Monograph"
         section_titles: "Used like a table of contents to indicate the order of attached FileSets' \"related sections\"."
         subject: "Keywords which describe what this monograph is about."
   sort_date: "Sort Date"
