+++ ruumba/.ruumba.yml
@@ -64,6 +64,7 @@ Metrics/BlockLength:
 Metrics/BlockNesting:
   Exclude:
     - 'app/views/monograph_catalog/_index_monograph.html.erb.rb'
+    - 'app/views/shared/_altmetric_dimensions.html.erb.rb'
 Metrics/LineLength:
   Enabled: false
