+++ spec/presenters/hyrax/monograph_presenter_spec.rb
@@ -17,6 +17,7 @@ RSpec.describe Hyrax::MonographPresenter do
     it do
       is_expected.to be_a CommonWorkPresenter
       is_expected.to be_a CitableLinkPresenter
+      is_expected.to be_a EditionPresenter
       is_expected.to be_a OpenUrlPresenter
       is_expected.to be_a TitlePresenter
       is_expected.to be_a FeaturedRepresentatives::MonographPresenter
@@ -554,172 +555,6 @@ RSpec.describe Hyrax::MonographPresenter do
       before { allow(mono_doc).to receive(:funder_display).and_return('Funder Display') }
 
       it { expect(presenter.funder_display?).to be true }
-    end
-  end
-
-  describe '#previous_edition_presenter' do
-    subject { presenter.previous_edition_presenter }
-
-    context 'no next edition link set' do
-      before { allow(mono_doc).to receive(:previous_edition).and_return(nil) }
-
-      it { expect(subject).to be nil }
-    end
-
-    context 'next edition link points to another Monograph with necessary fields set' do
-      let(:previous_edition_mono_doc) { ::SolrDocument.new(id: '000000000', has_model_ssim: ['Monograph'],
-                                                           edition_name_tesim: ['Second Edition'],
-                                                           isbn_tesim: ['978-0-472-55555-3 (ebook)'],
-                                                           doi_ssim: ['10.3998/mpub.5555555']) }
-
-      before do
-        allow(mono_doc).to receive(:previous_edition).and_return('https://test.host/concern/monographs/000000000')
-        # SolrService.add takes hashes not docs!
-        ActiveFedora::SolrService.add([previous_edition_mono_doc.to_h])
-        ActiveFedora::SolrService.commit
-      end
-
-      context 'user cannot read the linked previous_edition Monograph' do
-        before { allow(ability).to receive(:can?).with(:read, '000000000').and_return(false) }
-
-        it 'does not instantiate a MonographPresenter' do
-          expect(subject).to be_nil
-        end
-      end
-
-      context 'user can read the linked previous_edition Monograph' do
-        before { allow(ability).to receive(:can?).with(:read, '000000000').and_return(true) }
-
-        it 'instantiates a MonographPresenter' do
-          expect(subject).to be_present
-          expect(subject).to be_an_instance_of(Hyrax::MonographPresenter)
-          expect(subject.edition_name).to eq 'Second Edition'
-          expect(subject.isbn_noformat).to eq ['978-0-472-55555-3']
-          expect(subject.doi_path).to eq '10.3998/mpub.5555555'
-        end
-      end
-    end
-  end
-
-  describe '#previous_edition_url' do
-    subject { presenter.previous_edition_url }
-
-    context 'no previous edition link set' do
-      before { allow(mono_doc).to receive(:previous_edition).and_return(nil) }
-      it { expect(subject).to be nil }
-    end
-
-    context 'a Fulcrum Monograph link' do
-      before do
-        allow(mono_doc).to receive(:previous_edition).and_return('https://test.host/concern/monographs/111111111')
-      end
-
-      context 'user cannot read the linked previous_edition Monograph' do
-        before { allow(ability).to receive(:can?).with(:read, '111111111').and_return(false) }
-        it { expect(subject).to eq(nil) }
-      end
-
-      context 'user can read the linked previous_edition Monograph' do
-        before { allow(ability).to receive(:can?).with(:read, '111111111').and_return(true) }
-        it { expect(subject).to eq('https://test.host/concern/monographs/111111111') }
-      end
-    end
-
-    context 'external links' do
-      context 'is not a url' do
-        before { allow(mono_doc).to receive(:previous_edition).and_return('stuff') }
-        it { expect(subject).to eq(nil) }
-      end
-
-      context 'is a url' do
-        before { allow(mono_doc).to receive(:previous_edition).and_return('https://stuff') }
-        it { expect(subject).to eq('https://stuff') }
-      end
-    end
-  end
-
-  describe '#next_edition_url' do
-    subject { presenter.next_edition_url }
-
-    context 'no next edition link set' do
-      before { allow(mono_doc).to receive(:next_edition).and_return(nil) }
-      it { expect(subject).to be nil }
-    end
-
-    context 'a Fulcrum Monograph link' do
-      before do
-        allow(mono_doc).to receive(:next_edition).and_return('https://test.host/concern/monographs/222222222')
-      end
-
-      context 'user cannot read the linked next_edition Monograph' do
-        before { allow(ability).to receive(:can?).with(:read, '222222222').and_return(false) }
-        it { expect(subject).to eq(nil) }
-      end
-
-      context 'user can read the linked next_edition Monograph' do
-        before { allow(ability).to receive(:can?).with(:read, '222222222').and_return(true) }
-        it { expect(subject).to eq('https://test.host/concern/monographs/222222222') }
-      end
-    end
-
-    context 'external links' do
-      context 'is not a url' do
-        before { allow(mono_doc).to receive(:next_edition).and_return('blah') }
-        it { expect(subject).to eq(nil) }
-      end
-
-      context 'is a url' do
-        before { allow(mono_doc).to receive(:next_edition).and_return('https://stuff') }
-        it { expect(subject).to eq('https://stuff') }
-      end
-    end
-  end
-
-  describe '#previous_edition_noid' do
-    subject { presenter.previous_edition_noid }
-
-    context 'no previous edition link set' do
-      before { allow(mono_doc).to receive(:previous_edition).and_return(nil) }
-      it { expect(subject).to be nil }
-    end
-
-    context 'not a Fulcrum link' do
-      before { allow(mono_doc).to receive(:previous_edition).and_return('http://example.com') }
-      it { expect(subject).to be nil }
-    end
-
-    context 'Fulcrum FileSet link' do
-      before { allow(mono_doc).to receive(:previous_edition).and_return('https://test.host/concern/file_sets/333333333') }
-      it { expect(subject).to be nil }
-    end
-
-    context 'not a Fulcrum link' do
-      before { allow(mono_doc).to receive(:previous_edition).and_return('https://test.host/concern/monographs/333333333') }
-      it { expect(subject).to eq('333333333') }
-    end
-  end
-
-  describe '#next_edition_noid' do
-    subject { presenter.next_edition_noid }
-
-    context 'no next edition link set' do
-      before { allow(mono_doc).to receive(:next_edition).and_return(nil) }
-      it { expect(subject).to be nil }
-    end
-
-    context 'not a Fulcrum link' do
-      before { allow(mono_doc).to receive(:next_edition).and_return('http://example.com') }
-      it { expect(subject).to be nil }
-    end
-
-    context 'Fulcrum FileSet link' do
-      before { allow(mono_doc).to receive(:next_edition).and_return('https://test.host/concern/file_sets/444444444') }
-      it { expect(subject).to be nil }
-    end
-
-    context 'not a Fulcrum link' do
-      before { allow(mono_doc).to receive(:next_edition).and_return('https://test.host/concern/monographs/444444444') }
-      it { expect(subject).to eq('444444444') }
     end
   end
 end
